sdk: 7.x refactor & namespaces cleanup
