+++ packages/sdk/src/api/index.ts
@@ -1,3 +1,1 @@
-export { Router } from './Router';
-export { TradeRouter } from './TradeRouter';
-export { WalletCtx } from './WalletCtx';
+export * from './PolkadotApi';
