+++ packages/sdk/src/client/index.ts
@@ -1,4 +1,3 @@
 export { AssetClient } from './AssetClient';
 export { BalanceClient } from './BalanceClient';
 export { FarmClient } from './FarmClient';
-export { PolkadotApiClient } from './PolkadotApi';
