+++ packages/sdk/src/index.ts
@@ -1,9 +1,10 @@
-export { Router, TradeRouter } from './api';
-export * from './pool';
-export * from './client';
-export * from './types';
-export * from './errors';
 export * from './consts';
+export * from './client';
+export * from './errors';
+export * from './pool';
+export * from './sor';
+export * from './types';
+
 export * from './utils/bignumber';
 export * from './utils/math';
 export * from './utils/json';
