+++ packages/sdk/src/pool/PoolFactory.ts
@@ -1,10 +1,10 @@
-import { LbpPool, LbpPoolBase } from './lbp/LbpPool';
-import { OmniPool, OmniPoolBase } from './omni/OmniPool';
-import { StableSwap, StableSwapBase } from './stable/StableSwap';
-import { XykPool } from './xyk/XykPool';
-import { AavePool } from './aave/AavePool';
+import { LbpPool, LbpPoolBase } from './lbp';
+import { OmniPool, OmniPoolBase } from './omni';
+import { StableSwap, StableSwapBase } from './stable';
+import { XykPool } from './xyk';
+import { AavePool } from './aave';
 
-import { Pool, PoolBase, PoolType } from '../types';
+import { Pool, PoolBase, PoolType } from './types';
 
 export class PoolFactory {
   static get(pool: PoolBase): Pool {
