+++ packages/sdk/src/pool/aave/AavePoolClient.ts
@@ -5,10 +5,10 @@ import { stringToU8a } from '@polkadot/util';
 
 import { decodeEventLog } from 'viem';
 
-import { AAVE_ABI } from './AaveAbi';
-import { AaveTradeExecutorPoolData } from './types';
-
 import { HYDRADX_SS58_PREFIX } from '../../consts';
+import { ERC20Mapping } from '../../utils/erc20';
+import { findNestedKey } from '../../utils/json';
+
 import {
   PoolBase,
   PoolType,
@@ -16,11 +16,11 @@ import {
   PoolFees,
   PoolToken,
   PoolPair,
-} from '../../types';
-import { ERC20Mapping } from '../../utils/erc20';
-import { findNestedKey } from '../../utils/json';
-
+} from '../types';
 import { PoolClient } from '../PoolClient';
+
+import { AAVE_ABI } from './AaveAbi';
+import { AaveTradeExecutorPoolData } from './types';
 
 const SYNC_MM_EVENTS = ['Supply', 'Withdraw', 'Repay', 'Borrow'];
