+++ packages/sdk/src/pool/index.ts
@@ -1,10 +1,11 @@
 export { CachingPoolService } from './CachingPoolService';
 export { PoolFactory } from './PoolFactory';
 export { PoolService } from './PoolService';
-export * from './PoolUtils';
 
 export * from './aave';
 export * from './xyk';
 export * from './omni';
 export * from './lbp';
 export * from './stable';
+
+export * from './types';
