+++ packages/sdk/src/sor/route/index.ts
@@ -1,3 +1,3 @@
-export { RouteSuggester } from './suggester';
 export { Bfs, Path } from './bfs';
 export { Edge, getNodesAndEdges } from './graph';
+export { RouteSuggester } from './suggester';
