+++ packages/sdk/src/sor/route/suggester.ts
@@ -1,7 +1,7 @@
-import { PoolBase } from '../types';
 import { Bfs, Node, Path } from './bfs';
-
 import { getNodesAndEdges, Edge } from './graph';
+
+import { PoolBase } from '../../pool';
 
 export class RouteSuggester {
   /**
@@ -21,13 +21,23 @@ export class RouteSuggester {
    * @param pools - substrate based pools
    * @returns all possible path proposals
    */
-  getProposals(tokenIn: string, tokenOut: string | null, pools: PoolBase[]): Edge[][] {
+  getProposals(
+    tokenIn: string,
+    tokenOut: string | null,
+    pools: PoolBase[]
+  ): Edge[][] {
     const nodeEdges = getNodesAndEdges(pools);
     const poolAssets = Object.keys(nodeEdges);
-    const possiblePairs: Edge[] = poolAssets.map((node) => nodeEdges[node]).flat();
+    const possiblePairs: Edge[] = poolAssets
+      .map((node) => nodeEdges[node])
+      .flat();
     const bfs = new Bfs();
     const bfsGraph = bfs.buildAndPopulateGraph(poolAssets, possiblePairs);
-    const possiblePaths = bfs.findPaths(bfsGraph, parseInt(tokenIn), tokenOut ? parseInt(tokenOut) : null);
+    const possiblePaths = bfs.findPaths(
+      bfsGraph,
+      parseInt(tokenIn),
+      tokenOut ? parseInt(tokenOut) : null
+    );
     return this.parsePaths(possiblePaths);
   }
