+++ packages/sdk/test/script/examples/router/getBestBuy.ts
@@ -1,22 +1,21 @@
 import { ApiPromise } from '@polkadot/api';
-import { PoolService, TradeRouter, ZERO } from '../../../../src';
+import { PoolService, TradeRouter, TradeUtils } from '../../../../src';
 
 import { PolkadotExecutor } from '../../PjsExecutor';
 import { ApiUrl } from '../../types';
 
-class GetBestBuyPriceExample extends PolkadotExecutor {
+class GetBestBuyExample extends PolkadotExecutor {
   async script(api: ApiPromise): Promise<any> {
     const poolService = new PoolService(api);
+    const txUtils = new TradeUtils(api);
+
     const router = new TradeRouter(poolService);
-    const bestBuy = await router.getBestBuy('69', '690', '10');
-    const transaction = bestBuy.toTx(ZERO);
+
+    const trade = await router.getBestBuy('69', '690', '10');
+    const transaction = txUtils.buildBuyTx(trade);
     console.log('Transaction hash: ' + transaction.hex);
-    return bestBuy;
+    return trade;
   }
 }
 
-new GetBestBuyPriceExample(
-  ApiUrl.HydraDx,
-  'Get best buy price HydraDX',
-  true
-).run();
+new GetBestBuyExample(ApiUrl.HydraDx, 'Get best buy price HydraDX', true).run();
